Correccion de error en prestamos

+++ src/app/components/authors-list/authors-list.component.ts
@@ -1,10 +1,13 @@
 import { ChangeDetectorRef, Component, EventEmitter, Output, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { MatTable, MatTableDataSource } from '@angular/material/table';
 import { Author } from 'src/app/models/author';
+import { Book } from 'src/app/models/book';
 import { AuthorService } from 'src/app/services/author.service';
+import { BooksService } from 'src/app/services/books.service';
 import { AddAuthorDialogComponent } from '../dialogs/add-author-dialog/add-author-dialog.component';
 import { EditAuthorDialogComponent } from '../dialogs/edit-author-dialog/edit-author-dialog.component';
 
@@ -21,22 +24,24 @@ export class AuthorsListComponent {
     new MatTableDataSource<Author>();
   displayedColumns: string[] = [
     'id',
-    'name',
-    'last_name',
+    'name',/* 
+    'last_name', */
     'action'
   ];
 
   constructor(
     private changeDetectorRefs: ChangeDetectorRef,
     private dialog: MatDialog,
-    private authorService: AuthorService
+    private authorService: AuthorService,
+    private bookService: BooksService,
+    private snackBar: MatSnackBar
   ) {}
 
   ngAfterViewInit(): void {
-    this.loadBeneficiaries();
+    this.loadAuthors();
   }
 
-  loadBeneficiaries(): void {
+  loadAuthors(): void {
     this.authorService.getAuthors().subscribe(
       (data: Author[]) => {
         this.dataSource.data = data;
@@ -72,7 +77,7 @@ export class AuthorsListComponent {
         this.authorService.addAuthor(authorData).subscribe(
           (createdAuthor: Author) => {
             console.log('Autor creado con éxito:', createdAuthor);
-            this.loadBeneficiaries();
+            this.loadAuthors();
           },
           error => {
             console.error('Error al crear el autor:', error);
@@ -93,9 +98,12 @@ export class AuthorsListComponent {
       // Si se proporciona un resultado, se editó el autor.
       this.authorService.updateAuthor(result).subscribe(
         () => {
-          this.loadBeneficiaries();
+          this.loadAuthors();
           // Edición exitosa, puedes realizar alguna acción adicional si es necesario
           console.log('Autor editado con éxito');
+          this.snackBar.open('Autor editado con éxito.', 'Cerrar', {
+            duration: 4000,
+          });
         },
         (error) => {
           console.error('Error al editar autor:', error);
@@ -107,17 +115,35 @@ export class AuthorsListComponent {
   }
 
   deleteAuthor(id: number): void {
-    this.authorService.deleteAuthor(id).subscribe(
-      () => {
-        // Eliminación exitosa, actualizar la lista de autores
-        this.loadBeneficiaries();
-        console.log('Autor eliminado con éxito');
+    // Primero, verifica si el autor tiene libros asociados
+    this.bookService.getBooksByAuthorId(id).subscribe(
+      (books: Book[]) => {
+        if (books.length > 0) {
+          // Si hay libros asociados, muestra un mensaje al usuario
+          console.log('No se puede eliminar este autor porque tiene libros asociados.');
+          this.snackBar.open('No se puede eliminar este autor porque tiene libros asociados.', 'Cerrar', {
+            duration: 4000,
+          });
+        } else {
+          // Si no hay libros asociados, procede con la eliminación del autor
+          this.authorService.deleteAuthor(id).subscribe(
+            () => {
+              // Eliminación exitosa, actualizar la lista de autores
+              this.loadAuthors();
+              console.log('Autor eliminado con éxito');
+            },
+            (error) => {
+              console.error('Error al eliminar autor:', error);
+            }
+          );
+        }
       },
       (error) => {
-        console.error('Error al eliminar autor:', error);
+        console.error('Error al obtener los libros del autor:', error);
       }
     );
   }
+  
 
   applyFilter(event: Event): void {
     const filterValue = (event.target as HTMLInputElement).value;
