+++ src/app/components/editorials-list/editorials-list.component.ts
@@ -1,9 +1,12 @@
 import { ChangeDetectorRef, Component, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatSort } from '@angular/material/sort';
 import { MatTable, MatTableDataSource } from '@angular/material/table';
+import { Book } from 'src/app/models/book';
 import { Editorial } from 'src/app/models/editorial';
+import { BooksService } from 'src/app/services/books.service';
 import { EditorialService } from 'src/app/services/editorial.service';
 import { AddEditorialDialogComponent } from '../dialogs/add-editorial-dialog/add-editorial-dialog.component';
 import { EditEditorialDialogComponent } from '../dialogs/edit-editorial-dialog/edit-editorial-dialog.component';
@@ -30,7 +33,9 @@ export class EditorialsListComponent {
   constructor(
     private changeDetectorRefs: ChangeDetectorRef,
     private dialog: MatDialog,
-    private editorialService: EditorialService
+    private editorialService: EditorialService,
+    private bookService: BooksService,
+    private snackBar: MatSnackBar
   ) { }
 
   ngAfterViewInit(): void {
@@ -95,6 +100,9 @@ export class EditorialsListComponent {
           this.loadEditorial();
           // Edición exitosa, puedes realizar alguna acción adicional si es necesario
           console.log('Editorial editado con éxito');
+          this.snackBar.open('Editorial editado con éxito.', 'Cerrar', {
+            duration: 4000,
+          });
         },
         (error) => {
           console.error('Error al editar editorial:', error);
@@ -105,14 +113,31 @@ export class EditorialsListComponent {
   }
 
   deleteEditorial(id: number): void {
-    this.editorialService.deleteEditorial(id).subscribe(
-      () => {
-        // Eliminación exitosa, actualizar la lista de editorials
-        this.loadEditorial();
-        console.log('Editorial eliminado con éxito');
+    // Primero, verifica si el autor tiene libros asociados
+    this.bookService.getBooksByEditorialId(id).subscribe(
+      (books: Book[]) => {
+        if (books.length > 0) {
+          // Si hay libros asociados, muestra un mensaje al usuario
+          console.log('No se puede eliminar este autor porque tiene libros asociados.');
+          this.snackBar.open('No se puede eliminar esta editorial porque tiene libros asociados.', 'Cerrar', {
+            duration: 4000,
+          });
+        } else {
+          // Si no hay libros asociados, procede con la eliminación del autor
+          this.editorialService.deleteEditorial(id).subscribe(
+            () => {
+              // Eliminación exitosa, actualizar la lista de autores
+              this.loadEditorial();
+              console.log('Autor eliminado con éxito');
+            },
+            (error) => {
+              console.error('Error al eliminar autor:', error);
+            }
+          );
+        }
       },
       (error) => {
-        console.error('Error al eliminar editorial:', error);
+        console.error('Error al obtener los libros del autor:', error);
       }
     );
   }
